requirements.txt
requests==2.31.0
streamlit==1.29.0
pandas==2.1.4
python-dotenv==1.0.0
schedule==1.2.0
pyyaml==6.0.1
